+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/Reports.kt
@@ -1,35 +1,42 @@
 package io.github.surpsg.deltacoverage.report
 
+import io.github.surpsg.deltacoverage.config.ReportsConfig
 import io.github.surpsg.deltacoverage.diff.CodeUpdateInfo
+import io.github.surpsg.deltacoverage.report.intellij.report.ReportBound
+import io.github.surpsg.deltacoverage.report.intellij.report.ReportPathStrategy
 import org.jacoco.report.check.Rule
-import java.nio.file.Path
+import java.io.File
 
 open class FullReport(
-    private val baseReportDir: Path,
-    val reports: Set<Report>
+    val jacocoReports: List<JacocoReport>
 ) {
-    fun resolveReportAbsolutePath(report: Report): Path {
-        return baseReportDir.resolve(report.reportDirName)
+    fun resolveReportAbsolutePath(jacocoReport: JacocoReport): File {
+        val reportPathStrategy = when (jacocoReport.reportType) {
+            ReportType.HTML -> ReportPathStrategy.Html(jacocoReport.reportsConfig)
+            ReportType.XML -> ReportPathStrategy.Xml(jacocoReport.reportsConfig)
+            ReportType.CSV -> ReportPathStrategy.Csv(jacocoReport.reportsConfig)
+        }
+        return reportPathStrategy.buildReportPath(jacocoReport.reportBound)
     }
 }
 
 class JacocoDeltaReport(
-    baseReportDir: Path,
-    reports: Set<Report>,
+    reports: List<JacocoReport>,
     val codeUpdateInfo: CodeUpdateInfo,
-    val violation: Violation
-) : FullReport(baseReportDir, reports)
+    val violations: Violations
+) : FullReport(reports)
 
-data class Report(
+data class JacocoReport(
     val reportType: ReportType,
-    val reportDirName: String
+    val reportBound: ReportBound,
+    val reportsConfig: ReportsConfig
 )
 
 enum class ReportType {
     HTML, XML, CSV
 }
 
-data class Violation(
+data class Violations(
     val failOnViolation: Boolean,
     val violationRules: List<Rule>
 )