+++ delta-coverage-gradle/src/functionalTest/kotlin/io/github/surpsg/deltacoverage/gradle/DeltaCoverageReportsTest.kt
@@ -5,12 +5,14 @@ import io.github.surpsg.deltacoverage.gradle.test.GradleRunnerInstance
 import io.github.surpsg.deltacoverage.gradle.test.ProjectFile
 import io.github.surpsg.deltacoverage.gradle.test.RestorableFile
 import io.github.surpsg.deltacoverage.gradle.test.RootProjectDir
-import org.assertj.core.api.Assertions.assertThat
+import io.kotest.assertions.assertSoftly
+import io.kotest.matchers.file.shouldBeADirectory
+import io.kotest.matchers.file.shouldBeAFile
+import io.kotest.matchers.file.shouldContainFile
+import io.kotest.matchers.file.shouldExist
 import org.gradle.testkit.runner.GradleRunner
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.CsvSource
 import java.io.File
 
 @GradlePluginTest(TestProjects.SINGLE_MODULE)
@@ -33,87 +35,52 @@ class DeltaCoverageReportsTest {
         buildFile.restoreOriginContent()
     }
 
-    @ParameterizedTest
-    @CsvSource(
-        value = [
-            "html,  html,       true",
-            "csv,   report.csv, false",
-            "xml,   report.xml, false"
-        ]
-    )
-    fun `delta-coverage should create single report type`(
-        reportToGenerate: String,
-        expectedReportFile: String,
-        isDirectory: Boolean
-    ) {
-        // setup
+    @Test
+    fun `delta-coverage should create all jacoco reports`() {
+        // GIVEN
         val baseReportDir = "build/custom/reports/dir/jacoco/"
         buildFile.file.appendText(
             """
             deltaCoverageReport {
                 diffSource.file.set('$diffFilePath')
-                reportConfiguration.baseReportDir.set('$baseReportDir')
-                reportConfiguration.$reportToGenerate.set(true)
-            }
-        """.trimIndent()
-        )
-
-        // run // assert
-        gradleRunner
-            .runDeltaCoverageTask()
-            .assertOutputContainsStrings("Fail on violations: false. Found violations: 0")
-
-        // and assert
-        val actualReport: File = rootProjectDir.resolve(baseReportDir)
-            .resolve("deltaCoverage")
-            .resolve(expectedReportFile)
-        assertThat(actualReport).exists()
-        assertThat(actualReport.isDirectory)
-            .`as`("isDirectory")
-            .isEqualTo(isDirectory)
-    }
-
-    @Test
-    fun `delta-coverage should create deltaCoverage dir and full coverage with html, csv and xml reports`() {
-        // setup
-        val baseReportDir = "build/custom/reports/dir/jacoco/"
-        buildFile.file.appendText(
-            """
-            
-            deltaCoverageReport {
-                diffSource {
-                    file.set('$diffFilePath')
-                }
-                coverageBinaryFiles = jacocoTestReport.executionData
-                classesDirs = jacocoTestReport.classDirectories
-                srcDirs = jacocoTestReport.sourceDirectories
-                
                 reports {
+                    baseReportDir.set("$baseReportDir")
                     html.set(true)
                     xml.set(true)
                     csv.set(true)
                     fullCoverageReport.set(true)
-                    baseReportDir.set('$baseReportDir')
                 }
             }
         """.trimIndent()
         )
 
-        // run // assert
-        gradleRunner.runDeltaCoverageTask()
+        // WHEN // THEN
+        gradleRunner
+            .runDeltaCoverageTask()
+            .apply { println(output) }
+            .assertOutputContainsStrings("Fail on violations: false. Found violations: 0")
 
-        // and assert
-        rootProjectDir.resolve(baseReportDir).apply {
-            assertAllReportsCreated(resolve("deltaCoverage"))
-            assertAllReportsCreated(resolve("fullReport"))
-        }
+        // AND THEN
+        val baseReportDirFile = rootProjectDir.resolve(baseReportDir).resolve("coverage-reports")
+        assertAllReportsCreated(baseReportDirFile.resolve("delta-coverage"))
+        assertAllReportsCreated(baseReportDirFile.resolve("full-coverage-report"))
     }
 
     private fun assertAllReportsCreated(baseReportDir: File) {
-        assertThat(baseReportDir.list()).containsExactlyInAnyOrder("report.xml", "report.csv", "html")
-        assertThat(baseReportDir.resolve("html").list())
-            .containsExactlyInAnyOrder(
-                *expectedHtmlReportFiles("com.java.test")
-            )
+        val htmlReportDir = baseReportDir.resolve("html")
+        assertSoftly(htmlReportDir) {
+            shouldExist()
+            shouldBeADirectory()
+            shouldContainFile("index.html")
+            shouldContainFile("com.java.test")
+        }
+        sequenceOf("report.xml", "report.csv")
+            .map { file -> baseReportDir.resolve(file) }
+            .forEach { reportFile ->
+                assertSoftly(reportFile) {
+                    shouldExist()
+                    shouldBeAFile()
+                }
+            }
     }
 }